third_party/WebKit/LayoutTests/bluetooth/readValue.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="bluetooth-helpers.js"></script>
<script>
test(function(t) { assert_true(window.testRunner instanceof Object); t.done(); },
     'window.testRunner is required for the following tests.');

// Generic Access Service.
var serviceUUID = '00001800-0000-1000-8000-00805f9b34fb';
// Device Name Characteristic. Belongs to Generic Access.
var includedCharacteristicUUID = '00002a00-0000-1000-8000-00805f9b34fb';
// Reconnection Address. Belongs to Generic Access.
var unreadableCharacteristicUUID = '00002a03-0000-1000-8000-00805f9b34fb';

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('ConnectableDeviceAdapter');
  return navigator.bluetooth.requestDevice({
    filters: [{services: [serviceUUID]}]
  }).then(function(device) {
    return device.connectGATT();
  }).then(function(gattServer) {
    return gattServer.getPrimaryService(serviceUUID);
  }).then(function(service) {
    return service.getCharacteristic(includedCharacteristicUUID);
  }).then(function(characteristic) {
    testRunner.setBluetoothMockDataSet('EmptyAdapter');
    return characteristic.readValue().then(function() {
      assert_unreached('Device went out of range, should fail.');
    }, function(e) {
      assert_equals(e.name, 'NetworkError');
      assert_equals(e.message, 'Bluetooth Device is no longer in range.');
    });
  });
}, 'Device goes out of range. Reject with NetworkError.');

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('ConnectableDeviceAdapter');
  return navigator.bluetooth.requestDevice({
    filters: [{services: [serviceUUID]}]
  }).then(function(device) {
    return device.connectGATT();
  }).then(function(gattServer) {
    return gattServer.getPrimaryService(serviceUUID);
  }).then(function(service) {
    return service.getCharacteristic(unreadableCharacteristicUUID);
  }).then(function(characteristic) {
    return characteristic.readValue().then(function() {
      assert_unreached('Characteristic is unreadable. Should fail.');
    }, function(e) {
      assert_equals(e.name, 'NotSupportedError');
    });
  });
}, 'Characteristic is not readable. Reject with NetworkError.');

// TODO(ortuno): Add a test for when a characterstics gets removed.
// http://crbug.com/499552

sequential_promise_test(function() {
  testRunner.setBluetoothMockDataSet('ConnectableDeviceAdapter');
  return navigator.bluetooth.requestDevice({
    filters: [{services: [serviceUUID]}]
  }).then(function(device) {
    return device.connectGATT();
  }).then(function(gattServer) {
    return gattServer.getPrimaryService(serviceUUID);
  }).then(function(service) {
    return service.getCharacteristic(includedCharacteristicUUID);
  }).then(function(characteristic) {
    return characteristic.readValue();
  }).then(function(value) {
    var decoder = new TextDecoder('utf-8');
    var value_str = decoder.decode(value);
    // TODO: Remove the .slice after https://crrev.com/1172853004 is in.
    assert_equals(value_str.slice(0, 17), 'Empty Mock Device');
  });
}, 'Request for characteristic. Should return right characteristic');
</script>
